dernier modif webcam ok et delete aussi

+++ functions/delete_img.php
@@ -1,22 +1,29 @@
 <?php
-
-if ($_POST['submit'] && $_POST['submit'] == "OK")
+include("image-core.php");
+function delete()
 {
-    if (!isset($_SESSION['id'])){
-	    header('Location:connexion.php');
-	    exit();
-    }
+
+if (!isset($_SESSION['id'])){
+header('Location:connexion.php');
+exit();
+}
 $user_ID = $_SESSION['id'];
-$img_ID = $_POST['image_ID'];
-    // $user_ID = 1;
-    // $img_ID = 1;
+$img_ID = $_POST['img_ID'];
+// $user_ID = 1;
+// $img_ID = 1;
+$req = $bdd->prepare('SELECT `user_ID` FROM `img` WHERE `img_ID`= ?');
+$req->execute(array($img_ID)) && $user_Id = $req->fetch();
+$user_id = $user_Id['user_ID']; //celui qui a poster img
+if($user_ID == $user_id)
+{
 $deletelike = $bdd->prepare("DELETE FROM `liked` WHERE `user_ID`= ? AND `img_ID`= ?");
 $deletelike->execute(array($user_ID, $img_ID));  
 
-$deletecontent = $bdd->prepare("DELETE FROM `img` WHERE `user_ID`= ? AND `img_ID`= ?");
-$deletecontent->execute(array($user_ID, $img_ID));  
+$deletecomment = $bdd->prepare("DELETE FROM `comment` WHERE `user_ID`= ? AND `img_ID`= ?");
+$deletecomment->execute(array($user_ID, $img_ID));  
 
-$deleteimg = $bdd->prepare("DELETE FROM `content` WHERE `user_ID`= ? AND `img_ID`= ?");
+$deleteimg = $bdd->prepare("DELETE FROM `img` WHERE `user_ID`= ? AND `img_ID`= ?");
 $deleteimg->execute(array($user_ID, $img_ID));  
+}
 }
 ?>
